chore(backend): cambios API, Application, Domain e Infrastructure (Ecuador/feature)
- Program.cs: registro servicios (CompanyMenu, LoteAveEngorde, LoteReproductoraAveEngorde, SeguimientoDiario, MovimientoPolloEngorde, SeguimientoAvesEngorde, LiquidacionTecnicaEcuador, IndicadorEcuador, etc.) y JsonDoubleConverter
- Controllers: Company, Farm, Lote, MovimientoAves, Produccion, ReporteContable, SeguimientoLoteLevante, TrasladoNavigation, etc.
- DTOs, Entities, Configurations, Context, Services y Migrations actualizados para multi-empresa/país y módulos Ecuador

+++ backend/src/ZooSanMarino.Application/DTOs/Produccion/SeguimientoItemDto.cs
@@ -39,10 +39,11 @@ public record SeguimientoItemDto(
     decimal? CoeficienteVariacion,
     string? ObservacionesPesaje,
     // Campos de agua (solo para Ecuador y Panamá)
-    decimal? ConsumoAguaDiario, // Consumo diario de agua en litros
-    decimal? ConsumoAguaPh, // Nivel de PH del agua
-    decimal? ConsumoAguaOrp, // Nivel de ORP (Oxidación-Reducción Potencial) del agua en mV
-    decimal? ConsumoAguaTemperatura // Temperatura del agua en °C
+    // NOTA: Usar double? para coincidir con double precision en PostgreSQL
+    double? ConsumoAguaDiario, // Consumo diario de agua en litros
+    double? ConsumoAguaPh, // Nivel de PH del agua
+    double? ConsumoAguaOrp, // Nivel de ORP (Oxidación-Reducción Potencial) del agua en mV
+    double? ConsumoAguaTemperatura // Temperatura del agua en °C
 );
